app/requirements.txt
wheel==0.37.1
bluepy==1.3.0
paho-mqtt==1.5.1
